refactor(QYOG-20): lib 폴더 이름을 utils로 변경
