+++ src/utils/filterHTMLAttribute/index.ts
